Abstract builder (articles only)
